implemented moduleInfo main menu
organised moduleInfo code a bit
