+++ src/main/java/seedu/duke/Ui.java
@@ -68,13 +68,21 @@ public class Ui {
     public static void printModuleInfoMessage() {
         System.out.println("Welcome to the module information menu ^~^\n"
             + "Please choose which action you would like to do and enter the number:\n"
-            + "[1] --- Add/View Module Description\n"
-            + "[2] --- Add/View Components and Their Weightages\n"
-            + "[3] --- View All Modules\n"
-            + "[4] --- Add a Review\n"
-            + "[5] --- View All Reviews\n"
-            + "[6] --- Delete modules\n"
-            + "[7] --- Exit to main menu\n");
+            + "[1]  --- Add New Module\n"
+            + "[2]  --- View a Module\n"
+            + "[3]  --- Add/View Components and Their Weightages\n"
+            + "[4]  --- Add Module's Modular Credits (MC)\n"
+            + "[5]  --- Add Module Grade\n"
+            + "[6]  --- View All Modules\n"
+            + "[7]  --- Add New Task\n"
+            + "[8]  --- Add Zoom Link\n"
+            + "[9]  --- Add a Review\n"
+            + "[10] --- View All Reviews\n"
+            + "[11] --- Delete Module\n"
+            + "[12] --- Delete Task\n"
+            + "[13] --- Delete Zoom Link\n"
+            + "[14] --- Delete Review\n"
+            + "[15] --- Exit to main menu\n");
     }
 
     public static void printTaskManagerMenu() {
@@ -203,130 +211,6 @@ public class Ui {
         System.out.println("You've deleted this: " + task.toString());
         System.out.println("Returning back to TaskManager menu now!");
         printHorizontalLine();
-    }
-
-    public static boolean printAllModulesIfNotEmpty(ArrayList<Module> modules) {
-        if (isEmptyModulesList(modules)) {
-            printReturnToModuleInfoMenuMessage();
-            return false;
-        }
-        System.out.println("Here are the modules in your Modules List:");
-        printHorizontalLine();
-        for (int i = 1; i <= modules.size(); ++i) {
-            System.out.println("[" + i + "] --- " + modules.get(i - 1).getName());
-        }
-        printHorizontalLine();
-        return true;
-    }
-
-    public static int readModuleNumberToBeDeleted(ArrayList<Module> modules) {
-        int moduleNumberInt = -1;
-        if (printAllModulesIfNotEmpty(modules)) {
-            printSelectModuleToDeleteMessage();
-            moduleNumberInt = readCommandToInt();
-            moduleNumberInt--;
-            return moduleNumberInt;
-        }
-        return moduleNumberInt;
-    }
-
-    public static void printDeletedModuleMessage(Module module) {
-        System.out.println("You've deleted this: " + module.getName());
-        System.out.println("NOTE: You are deleting your module description\n"
-            + module.getDescription());
-        if (!module.getReview().trim().isEmpty()) {
-            System.out.println("NOTE: You are deleting your review\n"
-                + module.getReview());
-        }
-        printHorizontalLine();
-    }
-
-    public static void printReviewMenu(ArrayList<Module> modules) {
-        if (isEmptyModulesList(modules)) {
-            printReturnToModuleInfoMenuMessage();
-            return;
-        }
-        printAllModulesIfNotEmpty(modules);
-    }
-
-    public static void printAllReviews(ArrayList<Module> modules) {
-        if (isEmptyModulesList(modules)) {
-            return;
-        }
-        printHorizontalLine();
-        for (Module module : modules) {
-            System.out.println("For " + module.getName() + ":");
-            if (module.getReview().equals("")) {
-                System.out.println("You have not reviewed this module yet.");
-            } else {
-                System.out.println(module.getReview());
-            }
-            printHorizontalLine();
-        }
-    }
-
-    public static boolean isEmptyModulesList(ArrayList<Module> modules) {
-        if (modules.isEmpty()) {
-            logger.log(Level.INFO, "You have not added any modules.");
-            return true;
-        }
-        return false;
-    }
-
-    public static String printAddReviewMessage(Module module) {
-        if (!module.getReview().equals("")) {
-            System.out.println("You already have added a review:");
-            System.out.println(module.getReview());
-            System.out.println("Would you like to replace this with another review? [Y/N]");
-            logger.log(Level.WARNING, "You will delete your old review. This cannot be undone.");
-            String command = readCommand();
-            if (readYN(command) == 0) {
-                System.out.println("Okay:) You still have the same review!");
-                printReturnToModuleInfoMenuMessage();
-                return module.getReview();
-            } else if (readYN(command) == 2) {
-                printReturnToModuleInfoMenuMessage();
-                return module.getReview();
-            }
-            assert readYN(command) == 1 : "readYN(command) should be 1 here";
-        }
-        System.out.println("After you finish your review, "
-            + "type '/end' to finish reviewing.");
-        System.out.println("Enter your review for " + module.getName() + " below: ");
-        return readReview();
-    }
-
-    public static int readYN(String command) {
-        if (command.equalsIgnoreCase("N")) {
-            return 0;
-        } else if (!command.equalsIgnoreCase("Y")) {
-            System.out.println("You did not enter a valid letter:(");
-            return 2;
-        }
-        return 1;
-    }
-
-    public static String readReview() {
-        StringBuilder review = new StringBuilder();
-        while (true) {
-            String input = Ui.readCommand();
-            review.append(input);
-            review.append("\n");
-            if (input.contains("/end")) {
-                break;
-            }
-        }
-        //drop everything after "/end"
-        String reviewString = review.toString().split("/end")[0];
-
-        printReviewAdded(reviewString);
-        return reviewString;
-    }
-
-    public static void printReviewAdded(String review) {
-        System.out.println("Woohoo~ Review added:");
-        System.out.println(review);
-        printReturnToModuleInfoMenuMessage();
     }
 
     public static void printSelectModuleToDeleteMessage() {
